+++ next.config.mjs
@@ -1,6 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   transpilePackages: ["next-sanity", "sanity"],
+  images: { remotePatterns: [{ protocol: "https", hostname: "cdn.sanity.io" }] },
 };
 
 export default nextConfig;